+++ psi4/driver/procrouting/dft/gga_functionals.py
@@ -211,7 +211,7 @@ funcs.append({
     "description": '   SOGGA11 Exchange-Correlation Functional\n',
 })
 
-# dashparam.dashcoeff now defines 'b97' (<-'b97-d') as Grimme's GGA
+# empirical_dispersion_resources.dashcoeff now defines 'b97' (<-'b97-d') as Grimme's GGA
 #   functional, so one could just define
 #       funcs.append({
 #            "name": "B97",
